+++ tests/test_component.py
@@ -1,6 +1,6 @@
 import pytest
 from asphalt.core import Context
-from motor.motor_asyncio import AsyncIOMotorClient, AsyncIOMotorReplicaSetClient
+from motor.motor_asyncio import AsyncIOMotorClient
 
 from asphalt.mongodb.component import MongoDBComponent
 
@@ -16,26 +16,7 @@ async def test_default_client(caplog):
     records.sort(key=lambda r: r.message)
     assert len(records) == 2
     assert records[0].message == ("Configured MongoDB client (default / ctx.mongo; "
-                                  "nodes='localhost:27017')")
-    assert records[1].message == 'MongoDB client (default) shut down'
-
-
-@pytest.mark.parametrize('address', [
-    'localhost:27017,remote.host:27019',
-    ['localhost:27017', 'remote.host:27019']
-], ids=['str', 'list'])
-@pytest.mark.asyncio
-async def test_replice_set_client(caplog, address):
-    """Test that a replica set is correctly configured and available on the context."""
-    async with Context() as context:
-        await MongoDBComponent(address=address, replicaSet='myReplicaSet').start(context)
-        assert isinstance(context.mongo, AsyncIOMotorReplicaSetClient)
-
-    records = [record for record in caplog.records if record.name == 'asphalt.mongodb.component']
-    records.sort(key=lambda r: r.message)
-    assert len(records) == 2
-    assert records[0].message == ("Configured MongoDB client (default / ctx.mongo; "
-                                  "nodes='localhost:27017,remote.host:27019')")
+                                  "host='localhost')")
     assert records[1].message == 'MongoDB client (default) shut down'
 
 
@@ -44,8 +25,8 @@ async def test_multiple_clients(caplog):
     """Test that a multiple connection configuration works as intended."""
     async with Context() as context:
         await MongoDBComponent(clients={
-            'db1': {'address': 'localhost:27018'},
-            'db2': {'address': '/tmp/mongodb.sock', 'ssl': True}
+            'db1': {'host': 'localhost:27018'},
+            'db2': {'host': '/tmp/mongodb.sock', 'ssl': True}
         }).start(context)
         assert isinstance(context.db1, AsyncIOMotorClient)
         assert isinstance(context.db2, AsyncIOMotorClient)
@@ -54,9 +35,9 @@ async def test_multiple_clients(caplog):
     records.sort(key=lambda r: r.message)
     assert len(records) == 4
     assert records[0].message == ("Configured MongoDB client (db1 / ctx.db1; "
-                                  "nodes='localhost:27018')")
+                                  "host='localhost:27018')")
     assert records[1].message == ("Configured MongoDB client (db2 / ctx.db2; "
-                                  "nodes='/tmp/mongodb.sock')")
+                                  "host='/tmp/mongodb.sock')")
     assert records[2].message == 'MongoDB client (db1) shut down'
     assert records[3].message == 'MongoDB client (db2) shut down'