+++ docs/versionhistory.rst
@@ -3,6 +3,12 @@ Version history
 
 This library adheres to `Semantic Versioning <http://semver.org/>`_.
 
+**2.0.0** (2016-11-05)
+
+- **BACKWARD INCOMPATIBLE** Upgraded to  Motor 1.0
+- **BACKWARD INCOMPATIBLE** The ``address`` option has been replaced with ``host``, which is now
+  directly passed to ``AsyncIOMotorEngine``
+
 **1.0.0** (2016-10-04)
 
 - Initial release